Unihertz - add this vendor
I've recently had this issue on a Unihertz device, and would like others to be
aware of the vendor-specific fix.

Unihertz is a PRC-based company that makes niche devices such as a super tiny
Android 10 phone, a physical keyboard phone, and a highly ruggedized Android 10
phone with a walkie-talkie attachment. The total devices in the wild are low,
but the use cases are niche enough that I believe most Unihertz phone users will
run into issues with the similarly niche apps they'll be using.