Rename _vendors-content/tecno/user.md to _vendors-content/tecnokg5k
We want to be maximally transparent about how we evaluate vendors and give them our negative score. So here it is:

    
Foreground services killed by default and the user cannot configure it otherwise
    
Foreground services killed by default and the user cannot reliably configure it otherwise
    
Foreground services killed by default, but the user can reliably configure it otherwise
    
Foreground services killed only when the user explicitly configures its phone to do so…
    
Particular foreground services are killed only when the user explicitly configure those services so.
We reserve the right to make little adjustments to the scoring to favour vendors who e.g. publish documentation on their non-standard battery savings or kill processes in a more gentle manner :).

This site is libre software, the code is available on GitHub.
Discussion and contributions welcomed!